+++ src/org/elixir_lang/mix/importWizard/conversion/ModuleSettings.kt
@@ -2,11 +2,12 @@ package org.elixir_lang.mix.importWizard.conversion
 
 import com.intellij.conversion.ConversionProcessor
 import com.intellij.conversion.ModuleSettings
+import org.elixir_lang.jps.shared.ElixirModuleTypeId.ELIXIR_MODULE_TYPE_ID
 import org.jetbrains.jps.model.serialization.java.JpsJavaModelSerializerExtension.EXCLUDE_OUTPUT_TAG
 
 class ModuleSettings : ConversionProcessor<ModuleSettings>() {
     override fun isConversionNeeded(moduleSettings: ModuleSettings) =
-            moduleSettings.moduleType == "ELIXIR_MODULE" &&
+            moduleSettings.moduleType == ELIXIR_MODULE_TYPE_ID &&
         parentElement(moduleSettings)?.getChild(EXCLUDE_OUTPUT_TAG) != null
 
     override fun process(moduleSettings: ModuleSettings) {